refactor: split FoldersView.svelte into SettingsView + tab components (#579 item 22)
FoldersView.svelte (1,311 lines) was actually the whole Settings screen with
6 unrelated tabs including a full theme builder, misnamed and oversized.
Split into a thin SettingsView router plus SettingsGeneral, SettingsFolders,
SettingsTools, SettingsThemes, and SettingsAbout, mirroring the
CollectionView/PlaylistsCollectionView router pattern. Extracted the shared
openExternalUrl helper. Theme preview/revert now uses onDestroy on
SettingsThemes instead of a settingsTab-comparison effect, since tab
switching naturally unmounts it.

Closes #612
