+++ src/lib/utils/openExternalUrl.ts
@@ -1,0 +1,8 @@
+export async function openExternalUrl(url: string) {
+  try {
+    const { openUrl } = await import("@tauri-apps/plugin-opener");
+    await openUrl(url);
+  } catch {
+    window.open(url, "_blank");
+  }
+}
